+++ lib/convert.ts
@@ -4,9 +4,9 @@
  * 
  * @packageDocumentation
  */
- import { RDFClass, RDFProperty, RDFIndividual, RDFPrefix, OntologyProperty, Vocab, Link, Example, global} from './common';
- import { RawVocabEntry, RawVocab, ValidationResults } from './common';
- import { validateWithSchema } from './schema';
+ import { RDFClass, RDFProperty, RDFIndividual, RDFPrefix, OntologyProperty, Vocab, Link, Status, Example} from './common';
+ import { RawVocabEntry, RawVocab, ValidationResults, global }                                             from './common';
+ import { validateWithSchema }                                                                             from './schema';
 
 /************************************************ Helper functions and constants **********************************/
 
@@ -51,6 +51,10 @@ const defaultPrefixes: RDFPrefix[] = [
     {
         prefix : "xsd",
         url    : "http://www.w3.org/2001/XMLSchema#"
+    },
+    {
+        prefix : "vs",
+        url    : "http://www.w3.org/2003/06/sw-vocab-status/ns#"
     }
 ];
 
@@ -150,6 +154,21 @@ function finalizeRawEntry(raw: RawVocabEntry): RawVocabEntry {
         return final;
     }
 
+    // The deprecation flag, as a separate value, is kept also for reasons of backward compatibility,
+    // but this makes the interpretation of the value(s) in the vocabulary a bit awkward. Later version
+    // (maybe even next version) would remove the deprecated flag from existing vocabularies, ie,
+    // switch to status, and all this will go away.
+
+    const {status, deprecated} = ((): {status: Status, deprecated: boolean} => {
+        if (raw.status !== undefined) {
+            return { status : raw.status, deprecated : raw.status === Status.deprecated }
+        } else if (raw.deprecated != undefined) {
+            return { status : raw.deprecated ? Status.deprecated: Status.stable, deprecated : raw.deprecated }
+        } else {
+            return { status: Status.stable, deprecated : false }
+        }
+    })();
+
     return {
         id          : (raw.id) ? raw.id : "Vocabulary definition error: no ID provided.",
         property    : raw.property,
@@ -158,7 +177,8 @@ function finalizeRawEntry(raw: RawVocabEntry): RawVocabEntry {
         upper_value : toArray(raw.upper_value),
         domain      : toArray(raw.domain),
         range       : toArray(raw.range),
-        deprecated  : (raw.deprecated === undefined) ? false : raw.deprecated,
+        deprecated  : deprecated,
+        status      : status,
         comment     : (raw.comment) ? cleanComment(raw.comment) : "",
         see_also    : toSeeAlso(raw.see_also),
         example     : toExample(raw.example),
@@ -176,10 +196,10 @@ function finalizeRawEntry(raw: RawVocabEntry): RawVocabEntry {
 function finalizeRawVocab(raw: RawVocab) : RawVocab {
     // Check whether the required entries (vocab and ontology) are present
     if (raw.vocab === undefined) {
-        throw("No 'vocab' section in the vocabulary specification.")
+        throw(new Error("No 'vocab' section in the vocabulary specification."));
     }
     if (raw.ontology === undefined) {
-        throw("No 'ontology' section in the vocabulary specification.")
+        throw(new Error("No 'ontology' section in the vocabulary specification."));
     }
 
     // It is perfectly fine if the vocab is not encoded as an array in YAML
@@ -213,7 +233,8 @@ function finalizeRawVocab(raw: RawVocab) : RawVocab {
 export function getData(vocab_source: string): Vocab {
     const validation_results: ValidationResults = validateWithSchema(vocab_source);
     if (validation_results.vocab === null) {
-        throw(validation_results.error);
+        const error = JSON.stringify(validation_results,null,4);
+        throw(new TypeError(`JSON Schema validation error`, {cause: error}));
     }
     const vocab: RawVocab = finalizeRawVocab(validation_results.vocab);
 
@@ -226,10 +247,10 @@ export function getData(vocab_source: string): Vocab {
     const prefixes: RDFPrefix[] = [
         ...vocab.vocab.map((raw: RawVocabEntry): RDFPrefix => {
             if (raw.id === undefined) {
-                throw "The vocabulary has no prefix"
+                throw(new Error("The vocabulary has no prefix"));
             }
             if (raw.value === undefined) {
-                throw "The vocabulary has no identifier"
+                throw(new Error("The vocabulary has no identifier"));
             }
             global.vocab_prefix = raw.id;
             global.vocab_url = raw.value;
@@ -263,27 +284,13 @@ export function getData(vocab_source: string): Vocab {
         ...defaultOntologyProperties,
     ];
 
-    // Get the classes. Note the special treatment for deprecated classes...
-    const classes: RDFClass[] = (vocab.class !== undefined) ? 
-        vocab.class.map((raw: RawVocabEntry): RDFClass => {
-            const types: string[] = (raw.deprecated) ? ["rdfs:Class", "owl:DeprecatedClass"] : ["rdfs:Class"];
-            return {
-                id         : raw.id,
-                type       : types,
-                label      : raw.label,
-                comment    : raw.comment,
-                deprecated : raw.deprecated,
-                subClassOf : raw.upper_value,
-                see_also   : raw.see_also,
-                example    : raw.example,
-            }
-        }) : [];
-
-    // Get the classes. Note the special treatment for deprecated properties, as well as 
+    // Get the properties. Note the special treatment for deprecated properties, as well as 
     // the extra owl types added depending on the range
     const properties: RDFProperty[] = (vocab.property !== undefined) ?
         vocab.property.map((raw: RawVocabEntry): RDFProperty => {
-            const types: string[] = (raw.deprecated) ? ["rdf:Property", "owl:DeprecatedProperty"] : ["rdfs:Property"];
+            const types: string[] = (raw.status === Status.deprecated) ? ["rdf:Property", "owl:DeprecatedProperty"] : ["rdfs:Property"];
+            // Calculate the number of entries in various categories
+            global.status_counter.add(raw.status);
             let range = raw.range;
             if (range && range.length > 0) {
                 if (range.length === 1 && (range[0].toUpperCase() === "IRI" || range[0].toUpperCase() === "URL")) {
@@ -306,12 +313,57 @@ export function getData(vocab_source: string): Vocab {
                 label         : raw.label,
                 comment       : raw.comment,
                 deprecated    : raw.deprecated,
+                status        : raw.status,
                 subPropertyOf : raw.upper_value,
                 see_also      : raw.see_also,
                 range         : range,
                 domain        : raw.domain,
                 example       : raw.example,
                 dataset       : raw.dataset,
+            }
+        }) : [];
+
+    // Get the classes. Note the special treatment for deprecated classes and the location of relevant domains and ranges
+    const classes: RDFClass[] = (vocab.class !== undefined) ? 
+        vocab.class.map((raw: RawVocabEntry): RDFClass => {
+            const types: string[] = (raw.status === Status.deprecated) ? ["rdfs:Class", "owl:DeprecatedClass"] : ["rdfs:Class"];
+            const range_of: string[] = [];
+            const domain_of: string[] = [];
+            const included_in_domain_of: string[] = [];
+            const includes_range_of: string[] = [];
+
+            // Calculate the number of entries in various categories
+            global.status_counter.add(raw.status);
+
+            // Get all domain/range cross references
+            for (const property of properties) {
+                const crossref = (refs: string[], single_ref: string[], multi_ref: string[]): void => {
+                    if( refs ) {
+                        // Remove the (possible) namespace reference from the CURIE
+                        const pure_refs = refs.map((range:string): string => {
+                            const terms = range.split(':');
+                            return terms.length === 1 ? range : terms[1]
+                        });
+                        if (pure_refs.length !== 0 && pure_refs.indexOf(raw.id) !== -1) {
+                            (pure_refs.length === 1 ? single_ref : multi_ref).push(property.id)
+                        }
+                    }
+                }
+                crossref(property.range, range_of, includes_range_of);
+                crossref(property.domain, domain_of, included_in_domain_of);
+            }
+
+            return {
+                id         : raw.id,
+                type       : types,
+                label      : raw.label,
+                comment    : raw.comment,
+                deprecated : raw.deprecated,
+                status     : raw.status,
+                subClassOf : raw.upper_value,
+                see_also   : raw.see_also,
+                example    : raw.example,
+                range_of, domain_of, included_in_domain_of, includes_range_of
             }
         }) : [];
 
@@ -323,6 +375,7 @@ export function getData(vocab_source: string): Vocab {
                 label         : raw.label,
                 comment       : raw.comment,
                 deprecated    : raw.deprecated,
+                status        : raw.status,
                 type          : (raw.upper_value !== undefined) ? raw.upper_value : [],
                 see_also      : raw.see_also,
                 example       : raw.example,