+++ lib/turtle.ts
@@ -4,7 +4,7 @@
  * 
  * @packageDocumentation
  */
-import { Vocab, global, RDFTerm, Link } from './common';
+import { Vocab, global, RDFTerm, Link, Status } from './common';
 
 /**
  * Generate the Turtle representation of the vocabulary.
@@ -43,7 +43,8 @@ export function toTurtle(vocab: Vocab): string {
     const commonFields = (entry: RDFTerm): void => {
         turtle += `    rdfs:label "${entry.label}" ;\n`;
         turtle += `    rdfs:comment """<div>${entry.comment}</div>"""^^rdf:HTML ;\n`;
-        turtle += `    rdfs:isDefinedBy cred: ;\n`;
+        turtle += `    rdfs:isDefinedBy <${global.vocab_url}> ;\n`;
+        turtle += `    vs:term_status "${entry.status}" ;\n`;
         if (entry.see_also && entry.see_also.length > 0) {
             const urls = entry.see_also.map( (link: Link): string => `<${link.url}>`).join(", ");
             turtle +=`    rdfs:seeAlso ${urls} ;\n`
@@ -63,7 +64,7 @@ export function toTurtle(vocab: Vocab): string {
     {
         // Block for the top level ontology entries
         turtle += "# Ontology definition\n";
-        turtle += "cred: a owl:Ontology ;\n";
+        turtle += `${global.vocab_prefix}: a owl:Ontology ;\n`;
         for (const ont of vocab.ontology_properties) {
             if (ont.property === 'dc:date') {
                 turtle += `    dc:date "${ont.value}"^^xsd:date ;\n`
@@ -82,7 +83,7 @@ export function toTurtle(vocab: Vocab): string {
         turtle += "# Class definitions\n"
         for (const cl of vocab.classes) {
             turtle += `${global.vocab_prefix}:${cl.id} a ${cl.type.join(", ")} ;\n`;
-            if (cl.deprecated) {
+            if (cl.status === Status.deprecated) {
                 turtle += `    owl:deprecated true ;\n`;
             }
             if (cl.subClassOf && cl.subClassOf.length > 0) {
@@ -97,7 +98,7 @@ export function toTurtle(vocab: Vocab): string {
         turtle += "# Property definitions\n"
         for (const prop of vocab.properties) {
             turtle += `${global.vocab_prefix}:${prop.id} a ${prop.type.join(", ")} ;\n`;
-            if (prop.deprecated) {
+            if (prop.status === Status.deprecated) {
                 turtle += `    owl:deprecated true ;\n`;
             }
             if (prop.subPropertyOf) {
@@ -117,7 +118,7 @@ export function toTurtle(vocab: Vocab): string {
         turtle += "# Definitions of individuals\n"
         for (const ind of vocab.individuals) {
             turtle += `${global.vocab_prefix}:${ind.id} a ${ind.type.join(", ")} ;\n`;
-            if (ind.deprecated) {
+            if (ind.status === Status.deprecated) {
                 turtle += `    owl:deprecated true ;\n`;
             }
             commonFields(ind);